move get_path() and prepare_path() functions to base class
