+++ mapchete/formats/default/geojson.py
@@ -164,35 +164,6 @@ class OutputData(base.OutputData):
             raise TypeError("invalid geometry type")
         return True
 
-    def get_path(self, tile):
-        """
-        Determine target file path.
-
-        Parameters
-        ----------
-        tile : ``BufferedTile``
-            must be member of output ``TilePyramid``
-
-        Returns
-        -------
-        path : string
-        """
-        return os.path.join(*[
-            self.path, str(tile.zoom), str(tile.row),
-            str(tile.col) + self.file_extension]
-        )
-
-    def prepare_path(self, tile):
-        """
-        Create directory and subdirectory if necessary.
-
-        Parameters
-        ----------
-        tile : ``BufferedTile``
-            must be member of output ``TilePyramid``
-        """
-        makedirs(os.path.dirname(self.get_path(tile)))
-
     def empty(self, process_tile=None):
         """
         Return empty data.
