+++ mapchete/formats/default/png_hillshade.py
@@ -165,35 +165,6 @@ class OutputData(base.OutputData):
         """
         return validate_values(config, [("path", str)])
 
-    def get_path(self, tile):
-        """
-        Determine target file path.
-
-        Parameters
-        ----------
-        tile : ``BufferedTile``
-            must be member of output ``TilePyramid``
-
-        Returns
-        -------
-        path : string
-        """
-        return os.path.join(*[
-            self.path, str(tile.zoom), str(tile.row),
-            str(tile.col) + self.file_extension]
-        )
-
-    def prepare_path(self, tile):
-        """
-        Create directory and subdirectory if necessary.
-
-        Parameters
-        ----------
-        tile : ``BufferedTile``
-            must be member of output ``TilePyramid``
-        """
-        makedirs(os.path.dirname(self.get_path(tile)))
-
     def profile(self, tile=None):
         """
         Create a metadata dictionary for rasterio.
